rafs: clear layout v5 code up
Prepare code base for v6 disk layout development.
