--- rafs/src/metadata/extended/mod.rs
@@ -1,5 +1,0 @@
-// Copyright 2021 Ant Group. All rights reserved.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod blob_table;
